+++ lib/src/navigator/navigator_tab_bar_view.dart
@@ -25,6 +25,7 @@ class NavigatorTabBarView extends StatefulWidget {
   const NavigatorTabBarView({
     super.key,
     required this.routeSettings,
+    this.keepIndex = false,
     this.childBuilder,
     this.controller,
     this.physics,
@@ -45,8 +46,9 @@ class NavigatorTabBarView extends StatefulWidget {
   /// list, as well as the [controller]'s [TabController.length].
   final List<RouteSettings> routeSettings;
 
-  final Widget Function(
-      BuildContext context, RouteSettings settings, Widget child)? childBuilder;
+  final bool keepIndex;
+
+  final Widget Function(BuildContext context, RouteSettings settings, Widget child)? childBuilder;
 
   /// How the page view should respond to user input.
   ///
@@ -176,16 +178,13 @@ class _NavigatorTabBarViewState extends State<NavigatorTabBarView> {
         return Future<void>.value();
       }
       _warpUnderwayCount += 1;
-      await _pageController.animateToPage(_currentIndex!,
-          duration: duration, curve: Curves.ease);
+      await _pageController.animateToPage(_currentIndex!, duration: duration, curve: Curves.ease);
       _warpUnderwayCount -= 1;
       return Future<void>.value();
     }
 
     assert((_currentIndex! - previousIndex).abs() > 1);
-    final initialPage = _currentIndex! > previousIndex
-        ? _currentIndex! - 1
-        : _currentIndex! + 1;
+    final initialPage = _currentIndex! > previousIndex ? _currentIndex! - 1 : _currentIndex! + 1;
     setState(() {
       _warpUnderwayCount += 1;
     });
@@ -196,8 +195,7 @@ class _NavigatorTabBarViewState extends State<NavigatorTabBarView> {
       return Future<void>.value();
     }
 
-    await _pageController.animateToPage(_currentIndex!,
-        duration: duration, curve: Curves.ease);
+    await _pageController.animateToPage(_currentIndex!, duration: duration, curve: Curves.ease);
     if (!mounted) {
       return Future<void>.value();
     }
@@ -217,20 +215,17 @@ class _NavigatorTabBarViewState extends State<NavigatorTabBarView> {
     }
 
     _warpUnderwayCount += 1;
-    if (notification is ScrollUpdateNotification &&
-        !_controller!.indexIsChanging) {
+    if (notification is ScrollUpdateNotification && !_controller!.indexIsChanging) {
       if ((_pageController.page! - _controller!.index).abs() > 1.0) {
         _controller!.index = _pageController.page!.round();
         _currentIndex = _controller!.index;
       }
-      _controller!.offset =
-          clampDouble(_pageController.page! - _controller!.index, -1.0, 1.0);
+      _controller!.offset = clampDouble(_pageController.page! - _controller!.index, -1.0, 1.0);
     } else if (notification is ScrollEndNotification) {
       _controller!.index = _pageController.page!.round();
       _currentIndex = _controller!.index;
       if (!_controller!.indexIsChanging) {
-        _controller!.offset =
-            clampDouble(_pageController.page! - _controller!.index, -1.0, 1.0);
+        _controller!.offset = clampDouble(_pageController.page! - _controller!.index, -1.0, 1.0);
       }
     }
     _warpUnderwayCount -= 1;
@@ -275,6 +270,7 @@ class _NavigatorTabBarViewState extends State<NavigatorTabBarView> {
             ? const PageScrollPhysics().applyTo(const ClampingScrollPhysics())
             : const PageScrollPhysics().applyTo(widget.physics),
         routeSettings: widget.routeSettings,
+        keepIndex: widget.keepIndex,
         childBuilder: widget.childBuilder,
       ),
     );